extract firebase into context, sign out button
