+++ src/Routes.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import App from './App';
 import AppWithContexts from './AppWithContexts';
 import Auth from './Auth';
+import { FirebaseContextProvider } from './contexts/FirebaseContext';
 import GlobalStyles from './global.styles';
 import Home from './Home';
 
@@ -11,13 +12,15 @@ const Routes: React.FC = () => {
   return (
     <>
       <GlobalStyles />
-      <Router>
-        <Switch>
-          <Route path="/s/:spaceId" component={AppWithContexts} />
-          <Route path="/home" component={Home} />
-          <Route path="/" component={Auth} />
-        </Switch>
-      </Router>
+      <FirebaseContextProvider>
+        <Router>
+          <Switch>
+            <Route path="/s/:spaceId" component={AppWithContexts} />
+            <Route path="/home" component={Home} />
+            <Route path="/" component={Auth} />
+          </Switch>
+        </Router>
+      </FirebaseContextProvider>
     </>
   );
 };
